Trait pallet_circuit::account_manager::AccountManager

pub trait AccountManager<Account, Balance, Hash, BlockNumber, AssetId> {
    fn get_charge_or_fail(
        charge_id: Hash
    ) -> Result<RequestCharge<Account, Balance, AssetId>, DispatchError>;
    fn no_charge_or_fail(charge_id: Hash) -> Result<(), DispatchError>;
    fn bump_contracts_registry_nonce() -> Result<Hash, DispatchError>;
    fn deposit(
        charge_id: Hash,
        payee: &Account,
        charge_fee: Balance,
        offered_reward: Balance,
        source: BenefitSource,
        role: CircuitRole,
        recipient: Option<Account>,
        maybe_asset_id: Option<AssetId>
    ) -> Result<(), DispatchError>;
    fn finalize(
        charge_id: Hash,
        outcome: Outcome,
        maybe_recipient: Option<Account>,
        maybe_actual_fees: Option<Balance>
    ) -> Result<(), DispatchError>;
    fn finalize_infallible(
        charge_id: Hash,
        outcome: Outcome,
        maybe_recipient: Option<Account>,
        maybe_actual_fees: Option<Balance>
    );
    fn on_collect_claimable(
        n: BlockNumber,
        r: RoundInfo<BlockNumber>
    ) -> Result<Vec<ClaimableArtifacts<Account, Balance>, Global>, DispatchError>;
    fn can_withdraw(
        beneficiary: &Account,
        amount: Balance,
        asset_id: Option<AssetId>
    ) -> bool;
    fn deposit_immediately(
        beneficiary: &Account,
        amount: Balance,
        asset_id: Option<AssetId>
    );
    fn withdraw_immediately(
        beneficiary: &Account,
        amount: Balance,
        asset_id: Option<AssetId>
    ) -> Result<(), DispatchError>;
}

Required Methods

fn get_charge_or_fail(
charge_id: Hash
) -> Result<RequestCharge<Account, Balance, AssetId>, DispatchError>

Lookup charge by Id and fail if not found

fn no_charge_or_fail(charge_id: Hash) -> Result<(), DispatchError>

Lookup charge by Id and fail if not found

fn bump_contracts_registry_nonce() -> Result<Hash, DispatchError>

Bump contracts registry nonce in Account Manager nonce state and return charge request Id

fn deposit(
charge_id: Hash,
payee: &Account,
charge_fee: Balance,
offered_reward: Balance,
source: BenefitSource,
role: CircuitRole,
recipient: Option<Account>,
maybe_asset_id: Option<AssetId>
) -> Result<(), DispatchError>

Send funds to a recipient via the escrow

fn finalize(
charge_id: Hash,
outcome: Outcome,
maybe_recipient: Option<Account>,
maybe_actual_fees: Option<Balance>
) -> Result<(), DispatchError>

Finalize a transaction, with an optional reason for failures

fn finalize_infallible(
charge_id: Hash,
outcome: Outcome,
maybe_recipient: Option<Account>,
maybe_actual_fees: Option<Balance>
)

Assert infallible finalize of a transaction if exists

fn on_collect_claimable(
n: BlockNumber,
r: RoundInfo<BlockNumber>
) -> Result<Vec<ClaimableArtifacts<Account, Balance>, Global>, DispatchError>

fn can_withdraw(
beneficiary: &Account,
amount: Balance,
asset_id: Option<AssetId>
) -> bool

fn deposit_immediately(
beneficiary: &Account,
amount: Balance,
asset_id: Option<AssetId>
)

fn withdraw_immediately(
beneficiary: &Account,
amount: Balance,
asset_id: Option<AssetId>
) -> Result<(), DispatchError>