Module pallet_circuit::bridges::messages::source_chain

Primitives of messages module, that are used on the source chain.

Structs

ForbidOutboundMessages

Structure that may be used in place of TargetHeaderChain, LaneMessageVerifier and MessageDeliveryAndDispatchPayment on chains, where outbound messages are forbidden.

RelayerRewards

Single relayer rewards.

Traits

LaneMessageVerifier

Lane message verifier.

MessageDeliveryAndDispatchPayment

Message delivery payment. It is called as a part of submit-message transaction. Transaction submitter is paying (in source chain tokens/assets) for:

TargetHeaderChain

Target chain API. Used by source chain to verify target chain proofs.

Type Definitions

RelayersRewards

Relayers rewards, grouped by relayer account id.

Sender

The sender of the message on the source chain.