Trait pallet_circuit::bridges::messages::target_chain::MessageDispatch

pub trait MessageDispatch<Fee> {
    type DispatchPayload: Decode;

    fn dispatch_weight(
        message: &DispatchMessage<Self::DispatchPayload, Fee>
    ) -> u64;
    fn dispatch(message: DispatchMessage<Self::DispatchPayload, Fee>);
}

Called when inbound message is received.

Required Associated Types

type DispatchPayload: Decode

Decoded message payload type. Valid message may contain invalid payload. In this case message is delivered, but dispatch fails. Therefore, two separate types of payload (opaque MessagePayload used in delivery and this DispatchPayload used in dispatch).

Required Methods

fn dispatch_weight(message: &DispatchMessage<Self::DispatchPayload, Fee>) -> u64

Estimate dispatch weight.

This function must: (1) be instant and (2) return correct upper bound of dispatch weight.

fn dispatch(message: DispatchMessage<Self::DispatchPayload, Fee>)

Called when inbound message is received.

It is up to the implementers of this trait to determine whether the message is invalid (i.e. improperly encoded, has too large weight, …) or not.