Module pallet_circuit::bridges::runtime

Primitives that may be used at (bridges) runtime level.

Structs

PreComputedSize

Pre-computed size.

StorageProofChecker

This struct is used to read storage values from a subset of a Merklized database. The “proof” is a subset of the nodes in the Merkle structure of the database, so that it provides authentication against a known Merkle root as well as the values in the database themselves.

Enums

SourceAccount

Type of accounts on the source chain.

StorageProofError

Constants

ACCOUNT_DERIVATION_PREFIX

A unique prefix for entropy when generating cross-chain account IDs.

BASILISK_CHAIN_ID

Basilisk

CALL_DISPATCH_MODULE_PREFIX

Call-dispatch module prefix.

CATALYST_CHAIN_ID

Catalyst

CIRCUIT_BRIDGE_INSTANCE

Bridge-with-Millau instance id.

CIRCUIT_CHAIN_ID

Bridge-with-Wococo instance id.

DALI_CHAIN_ID

Dali

DOLPHIN_CHAIN_ID

Dolphin

GATEWAY_BRIDGE_INSTANCE

Bridge-with-Rialto instance id.

GATEWAY_CHAIN_ID

Bridge-with-Rococo instance id.

GENSHIRO_CHAIN_ID

Genshiro

KUSAMA_CHAIN_ID

Bridge-with-Kusama instance id.

MILLAU_CHAIN_ID

Bridge-with-Millau instance id.

NO_INSTANCE_ID

Use this when something must be shared among all instances.

PANGOLIN_CHAIN_ID

Pangolin

POLKADOT_CHAIN_ID

Bridge-with-Polkadot instance id.

RIALTO_CHAIN_ID

Bridge-with-Rialto instance id.

ROCFINITY_CHAIN_ID

Rocfinity

ROCOCO_CHAIN_ID

Bridge-with-Rococo instance id.

ROCOCO_ENCOINTER_CHAIN_ID

Encointer

ROOT_ACCOUNT_DERIVATION_PREFIX

A unique prefix for entropy when generating a cross-chain account ID for the Root account.

SNOWBLINK_CHAIN_ID

Snowbridge

SOONSOCIAL_CHAIN_ID

SoonSocial

WOCOCO_CHAIN_ID

Bridge-with-Wococo instance id.

Traits

Chain

Minimal Substrate-based chain representation that may be used from no_std environment.

Size

Anything that has size.

Functions

craft_valid_storage_proof

Return valid storage proof and state root.

derive_account_id

Derive an account ID from a foreign account ID.

derive_relayer_fund_account_id

Derive the account ID of the shared relayer fund account.

Type Definitions

BlockNumberOf

Block number used by the chain.

ChainId

Unique identifier of the chain.

HashOf

Hash type used by the chain.

HasherOf

Hasher type used by the chain.

HeaderOf

Header type used by the chain.