Function pallet_circuit::bridges::test_utils::make_justification_for_header

pub fn make_justification_for_header<H>(
    params: JustificationGeneratorParams<H>
) -> GrandpaJustification<H>where
    H: Header,

Generate justifications in a way where we are able to tune the number of pre-commits and vote ancestries which are included in the justification.

This is useful for benchmarkings where we want to generate valid justifications with a specific number of pre-commits (tuned with the number of “authorities”) and/or a specific number of vote ancestries (tuned with the “votes” parameter).

Note: This needs at least three authorities or else the verifier will complain about being given an invalid commit.