Module pallet_circuit::executors

Structs

Bond

Generic type describing either an executor’s self-bond or a staker’s bond.

CancelledScheduledStakingRequest

Represents a cancelled scheduled request for emitting an event.

CandidateBondLessRequest

Request scheduled to change the executor candidate’s self-bond.

ExecutorInfo

Executor configuration information.

ExecutorSnapshot

Snapshot of collator state at the start of the round for which they are selected

ExecutorsMock

Fixtures

Protocol enforced thresholds and delays for staking.

NominatedStake

ScheduledConfigurationRequest

Represents a scheduled request for an executor configuration change. The request is executable if the provided RoundIndex is achieved.

ScheduledStakingRequest

Represents a scheduled request that define a StakingAction. The request is executable iff the provided RoundIndex is achieved.

Enums

CapacityStatus

Capacity status for top or bottom stakes.

ExecutorStatus

The activity status of the executor

StakeAdjust

Staker’s bond adjustment - used with locks.

StakerAdded

Convey relevant information describing if a delegator was added to the top or bottom Stakes added to the top yield a new total

StakerStatus

The activity status of the staker.

StakingAction

An action that can be performed upon a stake

Constants

EXECUTOR_LOCK_ID

STAKER_LOCK_ID

Traits

Executors