Module pallet_circuit::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use crate::weights::WeightInfo;

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

Structs

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

Events for the pallet.

Traits

Config

This pallet’s configuration trait

Type Definitions

EscrowBalance

FullSideEffects

Current Circuit’s context of active full side effects (requested + confirmation proofs) Lifecycle tips: FSX entries are created at the time of Xtx submission, where still uncertain whether Xtx will be accepted for execution (picked up in the bidding process).

LocalXtxStates

LocalXtxStates stores the map of LocalState - additional state to be used to communicate between SFX that belong to the same Xtx

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

PendingSFXBids

Temporary bids for SFX executions. Cleaned out each Config::BidsInterval, where are moved from PendingSFXBids to FSX::accepted_bids

PendingXtxBidsTimeoutsMap

Temporary bids for SFX executions. Cleaned out each Config::BidsInterval, where are moved from PendingSFXBids to FSX::accepted_bids

PendingXtxTimeoutsMap

Current Circuit’s context of active Xtx used for the on_initialize clock to discover the ones pending for execution too long, that eventually need to be killed

SFX2XTXLinksMap

Links mapping SFX 2 XTX

XExecSignals

Current Circuit’s context of all accepted for execution cross-chain transactions.