Trait pallet_circuit::pallet::Config

pub trait Config: Config {

    type SelfAccountId: Get<Self::AccountId>;
    type SelfGatewayId: Get<[u8; 4]>;
    type SelfParaId: Get<u32>;
    type XtxTimeoutDefault: Get<Self::BlockNumber>;
    type XtxTimeoutCheckInterval: Get<Self::BlockNumber>;
    type SFXBiddingPeriod: Get<Self::BlockNumber>;
    type DeletionQueueLimit: Get<u32>;
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Call: Parameter + Dispatchable<Origin = Self::Origin> + GetDispatchInfo + From<Call<Self>> + From<Call<Self>>;
    type WeightInfo: WeightInfo;
    type Balances: Inspect<Self::AccountId> + Mutate<Self::AccountId>;
    type Xdns: Xdns<Self>;
    type XBIPortal: XBIPortal<Self>;
    type XBIPromise: XBIPromise<Self, <Self as Config>::Call>;
    type Executors: Executors<Self, <<Self::Escrowed as EscrowTrait<Self>>::Currency as Currency<Self::AccountId>>::Balance>;
    type AccountManager: AccountManager<Self::AccountId, <<Self::Escrowed as EscrowTrait<Self>>::Currency as Currency<Self::AccountId>>::Balance, Self::Hash, Self::BlockNumber, u32>;
    type Escrowed: EscrowTrait<Self>;
    type Portal: Portal<Self>;
    type SignalQueueDepth: Get<u32>;

}

This pallet’s configuration trait

Required Associated Types

type SelfAccountId: Get<Self::AccountId>

The Circuit’s account id

type SelfGatewayId: Get<[u8; 4]>

The Circuit’s self gateway id

type SelfParaId: Get<u32>

The Circuit’s self parachain id

type XtxTimeoutDefault: Get<Self::BlockNumber>

The Circuit’s Default Xtx timeout

type XtxTimeoutCheckInterval: Get<Self::BlockNumber>

The Circuit’s Xtx timeout check interval

type SFXBiddingPeriod: Get<Self::BlockNumber>

The Circuit’s SFX Bidding Period

type DeletionQueueLimit: Get<u32>

The Circuit’s deletion queue limit - preventing potential delay when queue is too long in on_initialize

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type Call: Parameter + Dispatchable<Origin = Self::Origin> + GetDispatchInfo + From<Call<Self>> + From<Call<Self>>

A dispatchable call.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type Balances: Inspect<Self::AccountId> + Mutate<Self::AccountId>

A type that provides inspection and mutation to some fungible assets

type Xdns: Xdns<Self>

A type that provides access to Xdns

type XBIPortal: XBIPortal<Self>

type XBIPromise: XBIPromise<Self, <Self as Config>::Call>

type Executors: Executors<Self, <<Self::Escrowed as EscrowTrait<Self>>::Currency as Currency<Self::AccountId>>::Balance>

type AccountManager: AccountManager<Self::AccountId, <<Self::Escrowed as EscrowTrait<Self>>::Currency as Currency<Self::AccountId>>::Balance, Self::Hash, Self::BlockNumber, u32>

A type that provides access to AccountManager

type Escrowed: EscrowTrait<Self>

A type that manages escrow, and therefore balances

type Portal: Portal<Self>

A type that gives access to the new portal functionality

type SignalQueueDepth: Get<u32>

The maximum number of signals that can be queued for handling.

When a signal from 3vm is requested, we add it to the queue to be handled by on_initialize

This allows us to process the highest priority and mitigate any race conditions from additional steps.

The reasons for limiting the queue depth are:

The queue is in storage in order to be persistent between blocks. We want to limit the amount of storage that can be consumed.

The queue is stored in a vector and needs to be decoded as a whole when reading it at the end of each block. Longer queues take more weight to decode and hence limit the amount of items that can be deleted per block.