Type Definition pallet_circuit::pallet::FullSideEffects

pub type FullSideEffects<T> = StorageMap<_GeneratedPrefixForStorageFullSideEffects<T>, Identity, XExecSignalId<T>, Vec<Vec<FullSideEffect<<T as Config>::AccountId, <T as Config>::BlockNumber, EscrowedBalanceOf<T, <T as Config>::Escrowed>>>>, OptionQuery>;

Current Circuit’s context of active full side effects (requested + confirmation proofs) Lifecycle tips: FSX entries are created at the time of Xtx submission, where still uncertain whether Xtx will be accepted for execution (picked up in the bidding process).

@Circuit::Requested: create FSX array without confirmations or bids

@Circuit::Bonded -> Ready: add bids to FSX

@Circuit::PendingExecution -> add more confirmations at receipt

If no bids have been received @Circuit::PendingBidding, FSX entries will stay - just without the Bid. The details on Xtx status might be played back by looking up with the SFX2XTXLinksMap