Module pallet_circuit::side_effect

Modules

interface

parser

Structs

ConfirmedSideEffect

FullSideEffect

HardenedSideEffect

SFXBid

All Executors from the active set can bid for SFX executions in order to claim the rewards (max_fee) set by users, ultimately competing against one another on the open market rules. In case bid goes on Optimistic SFX, Executor will also have their bonded stake reserve to insure other Optimistic Executors co-executing given Xtx with their bonded collateral (reserved_bond) Their balance

SideEffect

SideEffectInterface

Enums

ConfirmationOutcome

Error

SecurityLvl

Constants

ADD_LIQUIDITY_SIDE_EFFECT_ID

ASSETS_TRANSFER_SIDE_EFFECT_ID

CALL_SIDE_EFFECT_ID

COMPOSABLE_CALL_SIDE_EFFECT_ID

DATA_SIDE_EFFECT_ID

EVM_CALL_SIDE_EFFECT_ID

ORML_TRANSFER_SIDE_EFFECT_ID

SWAP_SIDE_EFFECT_ID

TRANSFER_SIDE_EFFECT_ID

WASM_CALL_SIDE_EFFECT_ID

Type Definitions

EventSignature

SideEffectId

SideEffectName

TargetId