Module pallet_circuit::state

Structs

LocalXtxCtx

XExecSignal

A composable cross-chain (X) transaction that has already been verified to be valid and submittable

Enums

CircuitRole

CircuitStatus

Status of Circuit storage items: Requested - default Requested -> Validated - successfully passed the validation Option<Validated -> PendingInsurance>: If there are some side effects that request insurance, the status will stay in PendingInsurance until all insurance deposits are committed Validated/PendingInsurance -> Ready - ready for relayers to pick up and start executing on targets Ready -> PendingExecution - at least one side effect has already been confirmed, but not all of them PendingExecution -> Finished - all of the side effects are confirmed, now awaiting for the decision about Revert/Commit Circuit::Apply -> called internally - based on the side effects confirmations decides: Ready -> Committed: All of the side effects have been successfully confirmed Ready -> Reverted: Some of the side effects failed and the Xtx was reverted

InsuranceEnact