Trait pallet_circuit::threevm::Remuneration

pub trait Remuneration<T, Balance>where
    T: Config,{
    fn try_remunerate<Module>(
        payee: &<T as Config>::AccountId,
        module: &Module
    ) -> Result<Remunerated<<T as Config>::Hash>, DispatchError>
    where
        Module: ModuleOperations<T, Balance>;
    fn try_remunerate_exact<Module>(
        payee: &<T as Config>::AccountId,
        amount: Balance,
        module: &Module
    ) -> Result<Remunerated<<T as Config>::Hash>, DispatchError>
    where
        Module: ModuleOperations<T, Balance>;
    fn try_finalize(
        ledger_id: <T as Config>::Hash,
        outcome: Outcome
    ) -> Result<(), DispatchError>;
}

Required Methods

fn try_remunerate<Module>(
payee: &<T as Config>::AccountId,
module: &Module
) -> Result<Remunerated<<T as Config>::Hash>, DispatchError>where
Module: ModuleOperations<T, Balance>,

Try to remunerate the fees from the given module

fn try_remunerate_exact<Module>(
payee: &<T as Config>::AccountId,
amount: Balance,
module: &Module
) -> Result<Remunerated<<T as Config>::Hash>, DispatchError>where
Module: ModuleOperations<T, Balance>,

Try to remunerate the fees from the given module with a custom balance

fn try_finalize(
ledger_id: <T as Config>::Hash,
outcome: Outcome
) -> Result<(), DispatchError>

Try to finalize a ledger item with an reason